Fixed wall facing!
So the dot-product trick doesn't work in some cases, depending on the
object's center.

Instead, assume all exported walls are back-facing.  If any polygon can
be found which includes both verts of the edge and is front-facing
(dot-product face normal and origin-to-camera), then we know the
exported wall is front-facing.  A non-front-facing polygon doesn't give
us an early-out.  Think of a cube's bottom.